Improve accuracy in pgfplot-output
In addcoords.m, accuracy ist set to %.5g, was %.4g. Lines nearly parallel to an
axis appeared jagged, e.g., for temperature, a small range, but large numbers. A
code to calculate the required accuracy from input is commented out. Add a bit
more comments at the begin of a .pgfplot-file.