First version of asym.m, untested.
Calculate the flow field for a stack of asymmetric membranes. Still needs the
field fullsolution (=~partialsolution) in the solverstruct.

Some restrictions apply: The liquid film and gaseous phase integrators in the
free space upstream of each membrane both expect an upstream state of a vapor,
due to their termination conditions. A liquid film in front of the membrane
expects the membrane front to be at z = 0. The gaseous phase integrator stops at
q1 = 0.