Remove bugs in asym, write q for two-phase flow
asym.m runs for a number of cases, not yet all are tested. In
asym>heatfluxcriterion, a q_vapor slightly smaller than qmin is allowed, to
succesfully conclude integration in the upstream free space. Correct a small, but
disastrous error in asym>integrate_vapliq. Calculate the heat flux for two-phase
flow, if writesolution is true. Code is moved to flowsetup>q2ph. Print sensible
information for two-phase flow in mstackstruct.printsolution.