Compute membrane properties from mass flow data
The membrane is percolated in reverse, from the support to the
separation layer. First the support then upstream - support - first
layer - downstream, and additional layers appended at the bottom.
See the next commit, tests/testdustygas/2ptsqdrtc.log for a log.
Also, Kahan's (2004) formula for solving a quadratic equation is used
here, but will not be used later.