Add R142b, as originally implemented by J. Rumpler
Some functions are added that work on vektorized input. The original
substance property functions may or may not accept vectorized input. For
R142b, all functions accept vectorized input.
The isobaric heat capacity of the liquid phase is now pressure
dependent. That goes not well with the function signature for the other
substances, which is substance.cpl(T).