Let heat capacity of liquid R142b not depend on pressure
The correlation for liquid heat capacity was developed and tested in the
range from 10 to 30 bar, and results in complex values for pressures
smaller than 0 bar. Hence, omit the dependency of the specific isobaric
heat capacity of liquid R142b on pressure. Evaluate cpl at 1 bar. There
is a deviation of less than 0.05% between 0 bar and 1 bar, with a small
maximum at 1 bar. At 0 bar, a number of terms would be omitted.

The omission of pressure dependency makes the function signature of cpl
comformant with the functions for cpl of the other substances.

The dynamic viscosity of the liquid R142b deviates from NIST data by 30%
at a temperature of 300K, becoming 60% at 350 K. However, other
experimental data is given in Landolt-Börnstein that supports this
deviation. See the Bachelor thesis of J. Rumpler, 2021.